requirements.txt
streamlit
requests
beautifulsoup4
sentence-transformers
faiss-cpu
mistralai
